experimental/spi/note_assistant_v2/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic[email]
requests
python-dotenv
grpcio
openai
anthropic
google-generativeai
google_auth_oauthlib
shotgun_api3
httpx
websockets
PyYAML>=6.0
openai-whisper

# CSV utilities
pandas>=2.0.0

# Image processing and OCR
easyocr
Pillow>=10.0.0
opencv-python>=4.5.0
